GraphQL-350: "The requested qty is not available" should be received instead of Internal server error
